newsletter-digest/requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
anthropic
feedparser
beautifulsoup4
pyyaml
python-dateutil
